rafactor: почистил закоммиценный код, вынес функции обработки текста в utils
