+++ src/components/main/main.js
@@ -1,5 +1,4 @@
 /* eslint-disable jsx-a11y/img-redundant-alt */
-//import {NavLink} from 'react-router-dom';
 import HorizontalScroll from 'react-scroll-horizontal';
 
 import './main.css';
