+++ src/components/seriesText/seriesText.js
@@ -1,37 +1,9 @@
 import './seriesText.css';
 
+import {getArrayText, chechMeaningfulPhrase} from '../../utils/utils';
+
 function SeriesText({data, isWide, openPopup}) {
-	const getArrayText = (str, phraseArr) => {
-		let resArr = [];
-
-		if (phraseArr) {
-			let currentText = str;
-
-			phraseArr.forEach((phrase) => {
-				const indexFinish = currentText.indexOf(phrase);
-				resArr.push(currentText.slice(0, indexFinish), phrase);
-				currentText = currentText.slice(indexFinish + phrase.length);
-			});
-
-			resArr.push(currentText);
-		} else {
-			resArr.push(str);
-		}
-
-		return resArr;
-	};
-
 	const finalArrayText = getArrayText(data.content, data.meaningfulPhrases);
-
-	const chechMeaningfulPhrase = (text, phraseArr) => {
-		if (phraseArr) {
-			return phraseArr.find((phrase) => {
-				return phrase === text;
-			});
-		} else {
-			return false;
-		}
-	};
 
 	const showReference = (meaningfulPhrase) => {
 		openPopup(meaningfulPhrase);
